webservice/requirements.txt
flask
cassandra-driver
uuid
datetime
redis
asyncio
asyncio-nats-client
asyncio-nats-streaming
